backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2
pydantic-settings==2.5.2
pydantic==2.9.2
python-dotenv==1.0.1
